moved effects to separate files. Made stop_the_light animate growth/regression
