+++ python/effects/examples.py
@@ -3,19 +3,8 @@ from hoe.animation_framework import Scene
 from hoe.animation_framework import EffectDefinition
 from hoe.animation_framework import Effect
 from hoe.animation_framework import FeedbackEffect
-
-
-class SolidBackground(Effect):
-    """Always return a color"""
-    def __init__(self, color=(255, 0, 0), layout=None, n_pixels=None):
-        Effect.__init__(self, layout, n_pixels)
-        self.color = color
-        print "Created with color", self.color
-
-    def next_frame(self, pixels, t, collaboration_state, osc_data):
-        for ii in range(self.n_pixels):
-            pixels[ii] = pixels[ii] if pixels[ii] else self.color
-
+import generic_effects
+import debugging_effects
 
 class SpatialStripesBackground(Effect):
     def next_frame(self, pixels, t, collaboration_state, osc_data):
@@ -71,23 +60,6 @@ class MovingDot(Effect):
         pixels[ii] = (spark_val, spark_val, spark_val)
 
 
-class AdjustableFillFromBottom(Effect):
-    def next_frame(self, pixels, t, collaboration_state, osc_data):
-        for ii, coord in enumerate(self.layout.pixels):
-            self.fill(pixels, t, coord, ii, osc_data)
-
-    def fill(self, pixels, time, coord, ii, osc_data):
-        if (not pixels[ii]) and "bottom_fill" in osc_data.faders and int(osc_data.faders["bottom_fill"]) > int(coord["row"]):
-            pixels[ii] = tuple([int(osc_data.faders[key]) for key in ['r','g','b']])
-
-
-class PrintOSC(Effect):
-    """A effect layer that just prints OSC info when it changes"""
-    def next_frame(self, pixels, t, collaboration_state, osc_data):
-        if osc_data.contains_change:
-            print "Frame's osc_data is", osc_data
-
-
 """osc_printing_effect = Effect("print osc", PrintOSC)
 red_effect=EffectDefinition("all red", DadJokes)
 spatial_background=EffectDefinition("spatial background", SpatialStripesBackground)
@@ -111,17 +83,17 @@ class SampleFeedbackEffect(FeedbackEffect):
     def compute_state(self, t, collaboration_state, osc_data):
         pass
 
-osc_printing_effect = PrintOSC()
+osc_printing_effect = debugging_effects.PrintOSC()
 spatial_background= SpatialStripesBackground()
-red_background = SolidBackground((255, 0, 0))
-blue_background = SolidBackground((0, 0, 255))
+red_background = generic_effects.SolidBackground((255, 0, 0))
+blue_background = generic_effects.SolidBackground((0, 0, 255))
 moving_dot = MovingDot()
 
 default_feedback_effect = SampleFeedbackEffect()
 
 __all__ = [
-    Scene("redgreenprinting", default_feedback_effect, osc_printing_effect, SolidBackground()),
-    Scene("blueredgreen", default_feedback_effect, AdjustableFillFromBottom(), red_background),
+    Scene("redgreenprinting", default_feedback_effect, osc_printing_effect, generic_effects.SolidBackground()),
+    Scene("blueredgreen", default_feedback_effect, generic_effects.AdjustableFillFromBottom(), red_background),
     Scene("bluewithdot", default_feedback_effect, moving_dot, blue_background),
     Scene("spatial scene", default_feedback_effect, spatial_background)
 ]
